Implement separate refinement methods structure
