+++ sxbp/sxbp_internal.h
@@ -110,6 +110,12 @@ sxbp_result_t sxbp_make_bitmap_for_bounds(
 // private, prints out a bitmap to the given stream, for debugging
 void sxbp_print_bitmap(sxbp_bitmap_t* bitmap, FILE* stream);
 
+// private, refines a figure using the 'shrink from end' method
+sxbp_result_t sxbp_refine_figure_shrink_from_end(
+    sxbp_figure_t* figure,
+    const sxbp_refine_figure_options_t* options
+);
+
 // private, macro to assist in 'return early if NULL pointer' error checks
 #define SXBP_RETURN_FAIL_IF_NULL(pointer) if (pointer == NULL) return SXBP_RESULT_FAIL_PRECONDITION
